Address review comments.
Fixes #745
