+++ changes/745-djpetti.rst
@@ -1,2 +1,1 @@
-- Added support for `FrozenSet` members in dataclasses.
-- Added a better error when attempting to use types from the `typing` module that are not supported by Pydantic.
+Added support for ``FrozenSet`` members in dataclasses, and a better error when attempting to use types from the ``typing`` module that are not supported by Pydantic.
